Refactoring Role Enum for user
