+++ src/main/java/demo/codeexample/company/CompanyDto.java
@@ -1,0 +1,13 @@
+package demo.codeexample.company;
+
+import lombok.Data;
+
+@Data
+public class CompanyDto {
+    private Long id;
+    private String companyName;
+    private String address;
+    private String email;
+
+
+}
