+++ src/main/java/demo/codeexample/logger/LoggerPort.java
@@ -1,7 +1,5 @@
 package demo.codeexample.logger;
 
-import demo.codeexample.logger.domain.LoggerAction;
-
 public interface LoggerPort {
     void log(LoggerAction action,
              Long userId,
