+++ src/main/java/demo/codeexample/logger/domain/LoggerEntity.java
@@ -1,5 +1,6 @@
 package demo.codeexample.logger.domain;
 
+import demo.codeexample.logger.LoggerAction;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
