+++ src/main/java/demo/codeexample/user/UserLookup.java
@@ -1,7 +1,5 @@
 package demo.codeexample.user;
 
-import demo.codeexample.user.domain.Role;
-
 import java.util.List;
 import java.util.Optional;
