+++ src/main/java/demo/codeexample/user/domain/UserRepository.java
@@ -1,5 +1,6 @@
 package demo.codeexample.user.domain;
 
+import demo.codeexample.user.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
